+++ typings/typings-installer.d.ts
@@ -1,3 +1,8 @@
 interface InstallOptions {
 	save?: boolean;
 }
+
+interface ITypingInstaller {
+    init(): Thenable<any> | Promise<any>;
+    displayName: string;
+}