+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.1.0
+Added support for Typings and maintained support for TSD.
+Added search functionality to both typings and TSD.
+
 ## 0.0.1
 Initial commit of typings-installer